refactor(game-engine): extract special-action handlers (S27.8.1)
Slice 1 de S27.8 (refactor du monolith `actions.ts`, 2706 lignes ->
cible <= 600 lignes). Premier deroule pour assoir le pattern
d'extraction : 6 handlers homogenes one-shot avec un cout de
dependance minimal.

Extraction
- Nouveau module `packages/game-engine/src/actions/special-actions.ts`
  (158l, exports nommes) regroupant :
  - handleHypnoticGaze
  - handleProjectileVomit
  - handleStab
  - handleChainsaw
  - handleBallAndChain
  - handleBombThrow
- Chaque handler suit la meme structure : gates `team` /
  `hasPlayerActed` / `can*`, execution via la mecanique dediee
  (`mechanics/<x>.ts`), `setPlayerAction` puis `checkPlayerTurnEnd`.
- Aucun couplage cross-handler ni reference a un helper interne de
  `actions.ts` : extraction triviale.

`actions.ts`
- Imports `handle*` depuis `./special-actions`. Anciens imports
  `execute*` retires (n'etaient utilises que par les handlers
  extraits ; les `can*` restent pour `getLegalMoves`).
- Suppression des 6 fonctions (~125 lignes brutes) remplacees par un
  comment block d'aiguillage.
- Taille passe de 2706 a 2595 lignes (-111 lignes nettes, ~4%).

Tests
- 4875 tests game-engine + 1755 tests server passent. Aucune
  regression : les smoke tests existants (chainsaw, hypnotic-gaze,
  bombardier, ball-and-chain, projectile-vomit, stab) couvraient deja
  les handlers, le contrat d'export (signatures et comportement)
  reste identique.

Reste pour atteindre la cible S27.8 (<= 600 lignes pour `actions.ts`,
<= 400 pour les autres) :
- handleMove / handleNormalMove / handleDodge / handleDodgeRoll /
  handleBallPickup
- handleBlock / handleBlockChoose / handlePushChoose /
  handleFollowUpChoose / handleRerollChoose / handleDumpOffChoose
- handleBlitz
- handlePass / handleHandoff / handleThrowTeamMate
- handleFoul
- handleEndTurn / handleEndPlayerTurn / handleLeap
